Move `persistence.MessageStore` into its own package.
