+++ src/ax/messagestore/store.go
@@ -1,20 +1,21 @@
-package persistence
+package messagestore
 
 import (
 	"context"
 
 	"github.com/jmalloc/ax/src/ax"
+	"github.com/jmalloc/ax/src/ax/persistence"
 )
 
-// MessageStore is an interface for persisting streams of messages.
-type MessageStore interface {
+// Store is an interface for manipulating persisted streams of messages.
+type Store interface {
 	// AppendMessages appends one or more messages to a named stream.
 	//
 	// offset is a zero-based index into the stream. An error is returned if
 	// offset is not the next unused offset in the stream.
 	AppendMessages(
 		ctx context.Context,
-		tx Tx,
+		tx persistence.Tx,
 		stream string,
 		offset uint64,
 		envs []ax.Envelope,
@@ -26,21 +27,8 @@ type MessageStore interface {
 	// stream does not exist.
 	OpenStream(
 		ctx context.Context,
-		tx Tx,
+		tx persistence.Tx,
 		stream string,
 		offset uint64,
-	) (MessageStream, bool, error)
-}
-
-// MessageStream is a stream of messages stored in a MessageStore.
-type MessageStream interface {
-	// Next advances the stream to the next message.
-	// It returns false if there are no more messages in the stream.
-	Next(ctx context.Context) (bool, error)
-
-	// Get returns the message at the current offset in the stream.
-	Get(ctx context.Context) (ax.Envelope, error)
-
-	// Close closes the stream.
-	Close() error
+	) (Stream, bool, error)
 }
